+++ package.json
@@ -49,6 +49,7 @@
     "panzoom": "^9.4.2",
     "register-service-worker": "^1.7.2",
     "semver": "^7.3.5",
+    "threads": "^1.6.4",
     "uuid": "^8.3.2",
     "vue": "^2.6.11",
     "vue-class-component": "^7.2.3",
@@ -105,6 +106,7 @@
     "shx": "^0.3.3",
     "standard-version": "^9.2.0",
     "string-replace-loader": "^3.0.1",
+    "threads-plugin": "^1.4.0",
     "typescript": "~3.9.3",
     "vue-cli-plugin-i18n": "^2.0.2",
     "vue-cli-plugin-vuetify": "^2.0.9",
