+++ src/workers/parseGcode.worker.ts
@@ -1,0 +1,4 @@
+import { expose } from 'threads/worker'
+import parseGcode from './parseGcode'
+
+expose(parseGcode)
